+++ src/main/java/com/reggarf/mods/create_better_motors/registry/CBMBlockEntityTypes.java
@@ -5,6 +5,8 @@ package com.reggarf.mods.create_better_motors.registry;
 
 
 import com.mrh0.createaddition.blocks.connector.base.ConnectorRenderer;
+import com.reggarf.mods.create_better_motors.content.alternator.blocks.AlternatorVisual;
+import com.reggarf.mods.create_better_motors.content.alternator.blocks.AlternatorRenderer;
 import com.reggarf.mods.create_better_motors.content.alternator.blocksentity.AndesiteAlternatorBlockEntity;
 import com.reggarf.mods.create_better_motors.content.alternator.blocksentity.BrassAlternatorBlockEntity;
 import com.reggarf.mods.create_better_motors.content.alternator.blocksentity.CopperAlternatorBlockEntity;
@@ -13,7 +15,6 @@ import com.reggarf.mods.create_better_motors.content.motors.blocksentity.*;
 import com.reggarf.mods.create_better_motors.content.multimeter.MultiMeterBlockEntity;
 import com.simibubi.create.AllPartialModels;
 import com.simibubi.create.content.kinetics.base.OrientedRotatingVisual;
-import com.simibubi.create.content.kinetics.base.ShaftRenderer;
 import com.simibubi.create.content.kinetics.base.ShaftVisual;
 import com.tterrag.registrate.util.entry.BlockEntityEntry;
 
@@ -32,7 +33,7 @@ public class CBMBlockEntityTypes {
             .blockEntity("starter_motor", StarterMotorBlockEntity::new)
             .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
             .validBlocks(CBMBlocks.STARTER_MOTOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
 
 
@@ -40,14 +41,14 @@ public class CBMBlockEntityTypes {
             .blockEntity("basic_motor", BasicMotorBlockEntity::new)
             .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
             .validBlocks(CBMBlocks.BASIC_MOTOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
 
     public static final BlockEntityEntry<HardenedMotorBlockEntity> HARDENED_MOTOR = REGISTRATE
             .blockEntity("hardened_motor", HardenedMotorBlockEntity::new)
             .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
             .validBlocks(CBMBlocks.HARDENED_MOTOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
 
 
@@ -55,55 +56,55 @@ public class CBMBlockEntityTypes {
             .blockEntity("blazing_motor", BlazingMotorBlockEntity::new)
             .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
             .validBlocks(CBMBlocks.BLAZING_MOTOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
 
     public static final BlockEntityEntry<NioticMotorBlockEntity> NIOTIC_MOTOR = REGISTRATE
             .blockEntity("niotic_motor", NioticMotorBlockEntity::new)
             .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
             .validBlocks(CBMBlocks.NIOTIC_MOTOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
 
     public static final BlockEntityEntry<SpritedMotorBlockEntity> SPIRITED_MOTOR = REGISTRATE
             .blockEntity("spirited_motor", SpritedMotorBlockEntity::new)
             .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
             .validBlocks(CBMBlocks.SPIRITED_MOTOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
 
     public static final BlockEntityEntry<NitroMotorBlockEntity> NITRO_MOTOR = REGISTRATE
             .blockEntity("nitro_motor", NitroMotorBlockEntity::new)
             .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
             .validBlocks(CBMBlocks.NITRO_MOTOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
 
 
     public static final BlockEntityEntry<AndesiteAlternatorBlockEntity> ANDESITE_ALTERNATOR = REGISTRATE
             .blockEntity("andesite_alternator", AndesiteAlternatorBlockEntity::new)
-            .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
+            .visual(() -> AlternatorVisual::new, false)
             .validBlocks(CBMBlocks.ANDESITE_ALTERNATOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
     public static final BlockEntityEntry<CopperAlternatorBlockEntity> COPPER_ALTERNATOR = REGISTRATE
             .blockEntity("copper_alternator", CopperAlternatorBlockEntity::new)
-            .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
+            .visual(() -> AlternatorVisual::new, false)
             .validBlocks(CBMBlocks.COPPER_ALTERNATOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
     public static final BlockEntityEntry<BrassAlternatorBlockEntity> BRASS_ALTERNATOR = REGISTRATE
             .blockEntity("brass_alternator", BrassAlternatorBlockEntity::new)
-            .visual(() -> OrientedRotatingVisual.of(AllPartialModels.SHAFT_HALF))
+            .visual(() -> AlternatorVisual::new, false)
             .validBlocks(CBMBlocks.BRASS_ALTERNATOR)
-            .renderer(() -> CBMHalfShaftRenderer::new)
+            .renderer(() -> AlternatorRenderer::new)
             .register();
 
     public static final BlockEntityEntry<MultiMeterBlockEntity> MULTIMETER = REGISTRATE
             .blockEntity("multimeter", MultiMeterBlockEntity::new)
             .visual(() -> ShaftVisual::new, false)
             .validBlocks(CBMBlocks.MULTIMETER)
-            .renderer(() -> ShaftRenderer::new)
+            .renderer(() -> com.simibubi.create.content.kinetics.base.ShaftRenderer::new)
             .register();
 
     public static void load() {  }